# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/example/wwq/controller/WwqCartController.java
@@ -1,6 +1,7 @@
 package com.example.wwq.controller;
 
 
+import com.example.wwq.kit.AuthorHelper;
 import com.example.wwq.kit.JSONResult;
 import com.example.wwq.service.IWwqCartService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,6 +31,9 @@ public class WwqCartController {
     @Autowired
     private IWwqCartService wwqCartService;
 
+    @Autowired
+    private AuthorHelper authorHelper;
+
     /**
      * 添加购物车
      * @param productId 商品id
@@ -40,11 +44,10 @@ public class WwqCartController {
     @ResponseBody
     public String addShopCart(HttpServletRequest req, @RequestParam(value="productId",required=true)String productId,
                               @RequestParam(value="buyNum",required=true)Integer buyNum){
-        String userId = "1";
-//        String userId = authorHelper.getUserId(req);
-//        if(userId == null){
-//            return JSONResult.init(301, "success", "user not login");
-//        }
+        String userId = authorHelper.getUserId(req);
+        if(userId == null){
+            return JSONResult.init(301, "success", "user not login");
+        }
         int ret = wwqCartService.addShopCart(productId, buyNum, userId);
         if(ret > 0){
             return JSONResult.init(200, "success", ret);
@@ -64,11 +67,10 @@ public class WwqCartController {
     @ResponseBody
     public String shopProductCartList(HttpServletRequest req,@RequestParam(value="pageNum",defaultValue="1",required=true) Integer pageNum,
                                       @RequestParam(value="pageSize",defaultValue="10",required=true) Integer pageSize){
-        String userId = "1";
-//        String userId = authorHelper.getUserId(req);
-//        if(userId == null){
-//            return JSONResult.init(301, "success", "user not login");
-//        }
+        String userId = authorHelper.getUserId(req);
+        if(userId == null){
+            return JSONResult.init(301, "success", "user not login");
+        }
         Map<String, Object> pageShopProductCart = wwqCartService.shopProductCartList(userId, pageNum, pageSize);
         return JSONResult.init(200, "success", pageShopProductCart);
     }
@@ -84,11 +86,10 @@ public class WwqCartController {
     @ResponseBody
     public String modifyShopCartProductNum(HttpServletRequest req,@RequestParam(value="shopCartId",required=true)String shopCartId,
                                            @RequestParam(value="productNum",required=true)Integer num){
-        String userId = "1";
-//        String userId = authorHelper.getUserId(req);
-//        if(userId == null){
-//            return JSONResult.init(301, "success", "user not login");
-//        }
+        String userId = authorHelper.getUserId(req);
+        if(userId == null){
+            return JSONResult.init(301, "success", "user not login");
+        }
         int ret = wwqCartService.modifyShopCartProductNum(shopCartId, num,userId);
         if(ret == 500){
             return JSONResult.init(500, "fail");
@@ -106,11 +107,10 @@ public class WwqCartController {
     @RequestMapping(value="/deleteShopCartProduct",produces="text/html;charset=UTF-8")
     @ResponseBody
     public String deleteShopCartProduct(HttpServletRequest req,@RequestParam(value="shopCartId",required=true)String shopCartId){
-        String userId = "1";
-//        String userId = authorHelper.getUserId(req);
-//        if(userId == null){
-//            return JSONResult.init(301, "success", "user not login");
-//        }
+        String userId = authorHelper.getUserId(req);
+        if(userId == null){
+            return JSONResult.init(301, "success", "user not login");
+        }
         int ret = wwqCartService.deleteShopCartProduct(shopCartId);
         return JSONResult.init(200, "success",ret);
     }
@@ -119,11 +119,10 @@ public class WwqCartController {
     @RequestMapping(value="/shopCartProductNum",produces="text/html;charset=UTF-8")
     @ResponseBody
     public String shopCartProductNum(HttpServletRequest req){
-        String userId = "1";
-//        String userId = authorHelper.getUserId(req);
-//        if(userId == null){
-//            return JSONResult.init(301, "success", "user not login");
-//        }
+        String userId = authorHelper.getUserId(req);
+        if(userId == null){
+            return JSONResult.init(301, "success", "user not login");
+        }
         Integer numList = wwqCartService.shopCartProductNum(userId);
         return JSONResult.init(200, "success", numList);
     }
